refactor: PiecesDto 생성
