+++ src/main/java/chess/web/dto/ChessStatusDto.java
@@ -1,21 +1,19 @@
 package chess.web.dto;
 
-import java.util.List;
-
 public class ChessStatusDto {
 
-    private final List<PieceDto> pieces;
+    private final PiecesDto piecesDto;
     private final double blackScore;
     private final double whiteScore;
 
-    public ChessStatusDto(List<PieceDto> pieces, double blackScore, double whiteScore) {
-        this.pieces = pieces;
+    public ChessStatusDto(PiecesDto piecesDto, double blackScore, double whiteScore) {
+        this.piecesDto = piecesDto;
         this.blackScore = blackScore;
         this.whiteScore = whiteScore;
     }
 
-    public List<PieceDto> getPieces() {
-        return pieces;
+    public PiecesDto getPiecesDto() {
+        return piecesDto;
     }
 
     public double getBlackScore() {
